LayoutTests/platform/ios/mathml/focus-visible-mouse-and-keyboard-expected.txt

This test checks that the MathML UA stylesheet applies outline via :focus-visible, not :focus.

1
+
2

PASS Mouse click on MathML element should not trigger UA focus outline
FAIL Keyboard focus on MathML element should trigger UA focus outline assert_equals: expected Element node <mn id="keyboard-target" tabindex="0">2</mn> but got Element node <mn id="mouse-target" tabindex="0">1</mn>
